Reproduce compiler bug
To reproduce: `roc check example/main.roc`

Note that if you swap out mainForHost for the currently
commented-out implementation of platform/main.roc, it no longer says
"The #UserApp module does not expose anything by the name main."
So, whether the app appears to expose `main` depends on the platform
somehow.

Note that it actually varies based on whether we call `main`.
If we call it from the platform, that error seems to go away.