Support http extension methods
Http allows any arbitrary methods defined in rfc2774 and implemented by
convention. These are implemented by the upstream reqwest library so
there is no reason why we can't also implement it.

This removes a source of a potential crash, which is on the path of user
submitted data (a http request).